+++ integrations/github/github/core/exporters/abstract_exporter.py
@@ -1,17 +1,18 @@
 from abc import ABC, abstractmethod
-from typing import Any
+from typing import Any, TypeVar, Generic
 from github.clients.http.base_client import AbstractGithubClient
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE, RAW_ITEM
 
+T = TypeVar("T", bound=AbstractGithubClient)
+AnyOption = TypeVar("AnyOption")
 
-class AbstractGithubExporter[T: AbstractGithubClient](ABC):
+
+class AbstractGithubExporter(Generic[T], ABC):
     def __init__(self, client: T) -> None:
         self.client = client
 
     @abstractmethod
-    async def get_resource[AnyOption: Any](self, options: AnyOption) -> RAW_ITEM: ...
+    async def get_resource(self, options: Any) -> RAW_ITEM: ...
 
     @abstractmethod
-    def get_paginated_resources[
-        AnyOption: Any
-    ](self, options: AnyOption) -> ASYNC_GENERATOR_RESYNC_TYPE: ...
+    def get_paginated_resources(self, options: Any) -> ASYNC_GENERATOR_RESYNC_TYPE: ...