+++ integrations/github/github/core/exporters/code_scanning_alert_exporter.py
@@ -1,5 +1,9 @@
 from github.core.exporters.abstract_exporter import AbstractGithubExporter
-from github.helpers.utils import enrich_with_repository, extract_repo_params
+from github.helpers.utils import (
+    IgnoredError,
+    enrich_with_repository,
+    extract_repo_params,
+)
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE, RAW_ITEM
 from loguru import logger
 from github.core.options import (
@@ -11,6 +15,14 @@ from github.clients.http.rest_client import GithubRestClient
 
 class RestCodeScanningAlertExporter(AbstractGithubExporter[GithubRestClient]):
 
+    def _get_ignored_errors(self) -> list[IgnoredError]:
+        return [
+            IgnoredError(
+                status=403,
+                message_prefix="Advanced Security must be enabled for this repository to use code scanning.",
+            ),
+        ]
+
     async def get_resource[
         ExporterOptionsT: SingleCodeScanningAlertOptions
     ](self, options: ExporterOptionsT) -> RAW_ITEM:
@@ -19,7 +31,9 @@ class RestCodeScanningAlertExporter(AbstractGithubExporter[GithubRestClient]):
         alert_number = params["alert_number"]
 
         endpoint = f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/code-scanning/alerts/{alert_number}"
-        response = await self.client.send_api_request(endpoint)
+        response = await self.client.send_api_request(
+            endpoint, ignored_errors=self._get_ignored_errors()
+        )
 
         logger.info(
             f"Fetched code scanning alert with number: {alert_number} for repo: {repo_name}"
@@ -37,6 +51,7 @@ class RestCodeScanningAlertExporter(AbstractGithubExporter[GithubRestClient]):
         async for alerts in self.client.send_paginated_request(
             f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/code-scanning/alerts",
             params,
+            ignored_errors=self._get_ignored_errors(),
         ):
             logger.info(
                 f"Fetched batch of {len(alerts)} code scanning alerts from repository {repo_name}"