+++ integrations/github/tests/github/clients/http/test_rest_client.py
@@ -129,6 +129,7 @@ class TestGithubRestClient:
                 method="GET",
                 params=expected_params,
                 return_full_response=True,
+                ignored_errors=None,
             )
 
     @pytest.mark.parametrize(